Keep the mirror sync out of the listing import ring
