+++ scripts/mutation/equivalent-mutants/shared-a-l.txt
@@ -28,7 +28,7 @@ src/shared/site-assignment.ts::renewalDeadlineBaseMs~00cywnq  ?? → ||   # pars
 src/shared/site-assignment.ts::assignSitesForEntries.site~0c03jq6  ?? → ||   # available.pop(): BuiltSite|undefined, and a BuiltSite object is always truthy
 src/shared/site-assignment.ts::sendSiteAssignmentEmail.config~0yvlh8h  ?? → ||   # getEmailConfig(): EmailConfig|null
 src/shared/site-assignment.ts::sendSiteAssignmentEmail.replyTo~195nl86  ?? → ||   # parseEmail(): ValidEmail|null, always truthy or null
-src/shared/db/listing-prices.ts::sourceRowStatements~0fyy4cr  ?? → ||   # unit_price is number|null, so the only falsy-non-null value is 0 and 0 ?? 0 === 0 || 0
+
 src/shared/ledger/project.ts::allBalances.add~0gr4ng6  ?? → ||   # allBalances: Map<string,number>.get; the only falsy-non-null number is 0, and 0 ?? 0 === 0 || 0
 src/shared/ledger/project.ts::balanceOf~128hnl9  ?? → ||   # balanceOf: same Map<string,number>.get fallback; 0 ?? 0 === 0 || 0
 src/shared/ledger/reconcile.ts::IDENTITY_FIELDS~0rfpt43  ?? → ||   # IDENTITY_FIELDS kind: string|undefined, only falsy-non-null is "" and "" ?? "" === "" || ""
